rename logs to logs/DEBRECATED
