+++ ui/src/api/logs_DEBRECATED/schema.ts
