+++ ui/src/pages/namespace/Instances/Detail/Main/Logs/index.tsx
@@ -20,7 +20,7 @@ import ScrollContainer from "./ScrollContainer";
 import { Toggle } from "~/design/Toggle";
 import { formatLogTime } from "~/util/helpers";
 import { useInstanceDetails } from "~/api/instances/query/details";
-import { useLogs } from "~/api/logs/query/get";
+import { useLogs } from "~/api/logs_DEBRECATED/query/get";
 import { useTranslation } from "react-i18next";
 
 const LogsPanel = () => {
