+++ ui/src/pages/namespace/Instances/Detail/store/type.ts
@@ -1,4 +1,4 @@
-import { FiltersObj } from "~/api/logs/query/get";
+import { FiltersObj } from "~/api/logs_DEBRECATED/query/get";
 
 export type StateType = {
   instanceId: string;
